requirements.txt
gradio
torch
transformers
Pillow
ftfy
regex
git+https://github.com/openai/CLIP.git
